+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.7.0] - 2024-04-17
+### Added
+- datatables for process and artifact view
+
+## [0.6.0] - 2024-04-10
 ### Fixed
+- read network structure from selection
 - structure and collection in URL query
 
 ## [0.5.0] - 2024-03-22